Add memory system call replacements.
These system calls are used for heap setup and so we can't yet remove
the musl wrappers, but this provides the kernel support that we'll need
when we do.

The mmap override now needs to call the function that implements the
read system call directly, rather than going via the system call layer.
While modifying that code, also fix a bug where we were modifying the
offset of the file descriptor during mmap.

After this change, all of the functions, including the `enclave_*` ones,
use the syscall convention for reporting errors: a positive result
indicates success, a negative result indicates failure, with the error
code negated.

This change also updates the callers of these functions to correctly
check the results.  One case, in lkl/setup.c, was checking the return
value of `lkl_sys_mmap` against `MAP_FIXED`, so would only detect errors
that returned `-EPERM` and not any other kind of failure.  This
particular place happened sufficiently early that failure probably did
not happen.